fix: [YSS-82] 카카오 이미지 url 요청 시 삭제 안되게 처리 (#67)

+++ yakssok/src/main/java/server/yakssok/global/infra/s3/S3FileApi.java
@@ -47,6 +47,8 @@ public class S3FileApi {
 	}
 
 	public void delete(String fileUrl) {
+		if (fileUrl == null) return;
+		if (isKakaoUrl(fileUrl)) return;
 		String key = extractKeyFromUrl(fileUrl);
 		deleteByKey(key);
 	}
@@ -62,11 +64,20 @@ public class S3FileApi {
 		}
 	}
 
-	private String extractKeyFromUrl(String fileUrl) {
-		if (fileUrl == null || !fileUrl.contains(".com/")) {
-			throw new S3FileException(ErrorCode.INVALID_INPUT_VALUE);
+
+	private String extractKeyFromUrl(String url) {
+		String bucket = aws.getS3().getBucket();
+		String region = aws.getRegion().getStaticRegion();
+
+		String base = "https://" + bucket + ".s3." + region + ".amazonaws.com/";
+		if (url.startsWith(base)) {
+			return url.substring(base.length());
 		}
-		return fileUrl.substring(fileUrl.indexOf(".com/") + 5);
+		throw new S3FileException(ErrorCode.FAILED_FILE_DELETE);
 	}
 
+	private boolean isKakaoUrl(String url) {
+		String u = url.toLowerCase();
+		return u.contains("kakaocdn.net") || u.contains("kakaocdn.com");
+	}
 }
